🚚[Rename] FilterCheckbox으로 네이밍 수정 후 공통 컴포넌트 폴더로 이동 #73
